Downgraded aws amplify version so that build for production passes
