+++ middleware/authToken.global.js
@@ -1,5 +1,5 @@
 import { useMainStore } from '@/store/index'
-//import auth from '@/services/auth.js'
+import auth from '@/services/auth.js'
 export default defineNuxtRouteMiddleware(async (to, from) => {
   if (process.server) return
   // Log the user out if their authorization token has expired
@@ -8,7 +8,7 @@ export default defineNuxtRouteMiddleware(async (to, from) => {
     const expDate = new Date(tokenExp)
     const currentDate = new Date()
     if (currentDate >= expDate) {
-      //await auth.logout()
+      await auth.logout()
     }
   }
 })
