+++ package.json
@@ -17,6 +17,8 @@
     "sass": "^1.66.1"
   },
   "dependencies": {
+    "@aws-amplify/auth": "4.5.8",
+    "@aws-amplify/core": "4.5.8",
     "@element-plus/nuxt": "^1.0.6",
     "@pinia-plugin-persistedstate/nuxt": "^1.2.0",
     "@pinia/nuxt": "^0.5.1",
